pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.3.1",
    "reportlab>=4.4.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "requests>=2.32.4",
    "paypalrestsdk>=1.13.3",
    "dropbox>=12.0.2",
    "google-api-python-client>=2.177.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "python-docx>=1.2.0",
    "authlib>=1.6.1",
    "python-dotenv>=1.1.1",
    "flask-migrate>=4.1.0",
    "docx>=0.2.4",
    "flask-wtf>=1.2.2",
    "pyjwt>=2.10.1",
    "razorpay>=1.4.2",
    "flask-cors>=6.0.1",
    "pypdf2>=3.0.1",
    "beautifulsoup4>=4.13.4",
]
